+++ ubelt/util_colors.py
@@ -21,6 +21,8 @@ force all ANSI text coloring functions to become no-ops. This defaults to the
 value of the ``bool(os.environ.get('NO_COLOR'))`` flag, which is compliant with
 [NoColor]_.
 
+New in 1.3.4: The :py:mod:`rich` backend was added as an alternative to pygments.
+
 
 Related work:
     https://github.com/Textualize/rich
@@ -42,24 +44,43 @@ import os
 NO_COLOR = bool(os.environ.get('NO_COLOR'))  # type: bool
 
 
-def highlight_code(text, lexer_name='python', **kwargs):
+def highlight_code(text, lexer_name='python', backend='pygments', **kwargs):
     """
     Highlights a block of text using ANSI tags based on language syntax.
 
     Args:
-        text (str): plain text to highlight
-        lexer_name (str): name of language. eg: python, docker, c++
-        **kwargs: passed to pygments.lexers.get_lexer_by_name
+        text (str):
+            Plain text to parse and highlight
+
+        lexer_name (str):
+            Name of language. eg: python, docker, c++.
+            For an exhaustive list see :func:`pygments.lexers.get_all_lexers`.
+            Defaults to "python".
+
+        backend (str):
+            Either "pygments" or "rich". Defaults to "pygments".
+
+        **kwargs:
+            If the backend is "pygments", passed to
+            pygments.lexers.get_lexer_by_name.
 
     Returns:
         str:
-            text - highlighted text If pygments is not installed, the plain
-            text is returned.
+            text - highlighted text if the requested backend is installed,
+            otherwise the plain text is returned unmodified.
 
     Example:
         >>> import ubelt as ub
         >>> text = 'import ubelt as ub; print(ub)'
         >>> new_text = ub.highlight_code(text)
+        >>> print(new_text)
+
+    Example:
+        >>> import ubelt as ub
+        >>> text = 'import ubelt as ub; print(ub)'
+        >>> new_text = ub.highlight_code(text, backend='pygments')
+        >>> print(new_text)
+        >>> new_text = ub.highlight_code(text, backend='rich')
         >>> print(new_text)
     """
     if NO_COLOR:
@@ -73,30 +94,59 @@ def highlight_code(text, lexer_name='python', **kwargs):
         'c': 'cpp',
     }.get(lexer_name.replace('.', ''), lexer_name)
     try:
-
-        if sys.platform.startswith('win32'):  # nocover
-            # Hack on win32 to support colored output
-            try:
-                import colorama
-                if not colorama.initialise.atexit_done:
-                    # Only init if it hasn't been done
-                    colorama.init()
-            except ImportError:
-                warnings.warn(
-                    'colorama is not installed, ansi colors may not work')
-
-        import pygments  # type: ignore
-        import pygments.lexers  # type: ignore
-        import pygments.formatters  # type: ignore
-        import pygments.formatters.terminal  # type: ignore
-
-        formatter = pygments.formatters.terminal.TerminalFormatter(bg='dark')
-        lexer = pygments.lexers.get_lexer_by_name(lexer_name, **kwargs)
-        new_text = pygments.highlight(text, lexer, formatter)
-
+        if backend == 'pygments':
+            new_text = _pygments_highlight(text, lexer_name, **kwargs)
+        elif backend == 'rich':
+            new_text = _rich_highlight(text, lexer_name, **kwargs)
+        else:
+            raise KeyError(backend)
     except ImportError:  # nocover
-        warnings.warn('pygments is not installed, code will not be highlighted')
+        warnings.warn(f'{backend} is not installed, code will not be highlighted')
         new_text = text
+    return new_text
+
+
+def _pygments_highlight(text, lexer_name, **kwargs):
+    """
+    Original pygments highlight logic
+    """
+    if sys.platform.startswith('win32'):  # nocover
+        # Hack on win32 to support colored output
+        try:
+            import colorama
+            if not colorama.initialise.atexit_done:
+                # Only init if it hasn't been done
+                colorama.init()
+        except ImportError:
+            warnings.warn(
+                'colorama is not installed, ansi colors may not work')
+
+    import pygments  # type: ignore
+    import pygments.lexers  # type: ignore
+    import pygments.formatters  # type: ignore
+    import pygments.formatters.terminal  # type: ignore
+
+    formatter = pygments.formatters.terminal.TerminalFormatter(bg='dark')
+    lexer = pygments.lexers.get_lexer_by_name(lexer_name, **kwargs)
+    new_text = pygments.highlight(text, lexer, formatter)
+    return new_text
+
+
+def _rich_highlight(text, lexer_name):  # nocover
+    """
+    Alternative rich-based highlighter
+
+    References:
+        https://github.com/Textualize/rich/discussions/3076
+    """
+    from rich.syntax import Syntax
+    from rich.console import Console
+    import io
+    syntax = Syntax(text, lexer_name, background_color='default')
+    stream = io.StringIO()
+    write_console = Console(file=stream, soft_wrap=True, color_system='standard')
+    write_console.print(syntax)
+    new_text = write_console.file.getvalue()
     return new_text
 
 
@@ -115,6 +165,9 @@ def color_text(text, color):
         str:
             text - colorized text.  If pygments is not installed plain text is
             returned.
+
+    SeeAlso:
+        https://rich.readthedocs.io/en/stable/markup.html
 
     Example:
         >>> text = 'raw text'
@@ -137,6 +190,7 @@ def color_text(text, color):
         >>> # xdoctest: +REQUIRES(module:pygments)
         >>> import pygments.console
         >>> import ubelt as ub
+        >>> # List available colors codes
         >>> known_colors = pygments.console.codes.keys()
         >>> for color in known_colors:
         ...     print(ub.color_text(color, color))